+++ src/__tests__/App.test.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import App from './App';
-import GalleryApp from './GalleryApp';
+import GalleryApp from '../components/GalleryApp';
 
 describe('App', () => {
     it('runs the GalleryApp', () => {
