+++ src/__tests__/Gallery.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import App from './App';
-import Thumbnail from './Thumbnail';
+
+import Gallery from '../components/Gallery';
 let image = {
     title: 'Cute Bunny',
     description: 'Isn\'t it fuzzy-wuzzy cutest thing you\'ve ever seen?',
@@ -10,9 +10,9 @@ let image = {
 };
 
 
-describe('Thumbnail', () => {
-    it('runs the Thumbnail component', () => {
-        const rendered = shallow(<Thumbnail image={image} />);
+describe('Gallery ', () => {
+    it('runs the Gallery component', () => {
+        const rendered = shallow(<Gallery image={image} />);
         const json = toJson(rendered);
         expect(json).toMatchSnapshot();
     })
