+++ src/__tests__/Thumbnail.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import App from './App';
-import ListImage from './ListImage';
+
+import Thumbnail from '../components/Thumbnail';
 let image = {
     title: 'Cute Bunny',
     description: 'Isn\'t it fuzzy-wuzzy cutest thing you\'ve ever seen?',
@@ -10,9 +10,9 @@ let image = {
 };
 
 
-describe('ListImage ', () => {
-    it('runs the ListImage component', () => {
-        const rendered = shallow(<ListImage image={image} />);
+describe('Thumbnail', () => {
+    it('runs the Thumbnail component', () => {
+        const rendered = shallow(<Thumbnail image={image} />);
         const json = toJson(rendered);
         expect(json).toMatchSnapshot();
     })
