+++ src/components/Thumbnail.js
