+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './App.css';
-import GalleryApp from './GalleryApp';
+import GalleryApp from './components/GalleryApp';
 
 
 ReactDOM.render(
